+++ public/data/articleData.json
@@ -1,0 +1,32 @@
+[
+  {
+    "user_id": "intarrrr",
+    "article_text": "집에 늦게 가는 것만이 능사가 아닙니다.",
+    "article_img": "profile1"
+  },
+  {
+    "user_id": "seulgi",
+    "article_text": "안녕하세요 레드벨벳 슬기입니다.",
+    "article_img": "슬기"
+  },
+  {
+    "user_id": "yeri",
+    "article_text": "안녕하세요 레드벨벳 예리입니다.",
+    "article_img": "예리"
+  },
+  {
+    "user_id": "wendy",
+    "article_text": "안녕하세요 레드벨벳 웬디입니다.",
+    "article_img": "웬디"
+  },
+  {
+    "user_id": "ChangChen",
+    "article_text": "안녕하세요 하얼빈 장첸입니다.",
+    "article_img": "장첸"
+  },
+  {
+    "user_id": "irene",
+    "article_text": "안녕하세요 레드벨벳 아이린입니다.",
+    "article_img": "아이린"
+  }
+]
